refactor(ai): extract AI_TASKS/AI_PROVIDERS/DEFAULT_MODELS to shared constants and dedup getProviderForTask
- Create utilities/ai-constants.js with AI_PROVIDERS, AI_TASKS, DEFAULT_MODELS
- Export getProviderForTask from ai-gateway (eliminates the duplicate in the controller)
- controllers/api/ai.js imports constants from ai-constants and getProviderForTask from gateway
- controllers/api/bienbot.js and utilities/bienbot-{memory-extractor,session-summarizer,external-data}.js import AI_TASKS directly from ai-constants
- Backward-compat exports on controllers/api/ai.js are retained (re-export from the new home)

Closes dc8d.7, dc8d.8
