+++ controllers/api/bienbot.js
@@ -37,7 +37,8 @@ const { extractText, validateDocument } = require('../../utilities/ai-document-u
 const { uploadWithPipeline, retrieveFile, resolveAndValidateLocalUploadPath } = require('../../utilities/upload-pipeline');
 const path = require('path');
 const { GatewayError } = require('../../utilities/ai-gateway');
-const { callProvider, getApiKey, getProviderForTask, AI_TASKS } = require('./ai');
+const { callProvider, getApiKey, getProviderForTask } = require('./ai');
+const { AI_TASKS } = require('../../utilities/ai-constants');
 const BienBotSession = require('../../models/bienbot-session');
 const affinityCache = require('../../utilities/affinity-cache');
 const { computeAndCacheAffinity } = require('../../utilities/hidden-signals');
