+++ utilities/bienbot-memory-extractor.js
@@ -12,7 +12,8 @@
 
 const logger = require('./backend-logger');
 const { executeAIRequest, GatewayError } = require('./ai-gateway');
-const { getApiKey, getProviderForTask, AI_TASKS } = require('../controllers/api/ai');
+const { getApiKey, getProviderForTask } = require('../controllers/api/ai');
+const { AI_TASKS } = require('./ai-constants');
 
 // ---------------------------------------------------------------------------
 // Constants
